Breaking up events components
